+++ ui/packages/shared/profile/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.19.109](https://github.com/parca-dev/parca/compare/@parca/profile@0.19.108...@parca/profile@0.19.109) (2026-01-07)
+
+**Note:** Version bump only for package @parca/profile
+
+## [0.19.108](https://github.com/parca-dev/parca/compare/@parca/profile@0.19.107...@parca/profile@0.19.108) (2026-01-05)
+
+**Note:** Version bump only for package @parca/profile
+
+## [0.19.107](https://github.com/parca-dev/parca/compare/@parca/profile@0.19.106...@parca/profile@0.19.107) (2025-12-22)
+
+**Note:** Version bump only for package @parca/profile
+
+## [0.19.106](https://github.com/parca-dev/parca/compare/@parca/profile@0.19.105...@parca/profile@0.19.106) (2025-12-18)
+
+**Note:** Version bump only for package @parca/profile
+
+## [0.19.105](https://github.com/parca-dev/parca/compare/@parca/profile@0.19.104...@parca/profile@0.19.105) (2025-12-18)
+
+**Note:** Version bump only for package @parca/profile
+
+## [0.19.104](https://github.com/parca-dev/parca/compare/@parca/profile@0.19.103...@parca/profile@0.19.104) (2025-12-17)
+
+**Note:** Version bump only for package @parca/profile
+
+## [0.19.103](https://github.com/parca-dev/parca/compare/@parca/profile@0.19.102...@parca/profile@0.19.103) (2025-12-17)
+
+**Note:** Version bump only for package @parca/profile
+
 ## [0.19.102](https://github.com/parca-dev/parca/compare/@parca/profile@0.19.101...@parca/profile@0.19.102) (2025-12-15)
 
 **Note:** Version bump only for package @parca/profile